Rename BoardButtons to SpaceButtons
